Implement drawing notes dynamically.
Implement a system for drawing notes dynamically from an array list,
which will have notes added to it during gameplay from information
read from a file in the future. These notes move down the screen,
and are removed from the array list after they've reached a certain
position past the bottom of the screen.