Fix red long notes not drawing correct image.
Fix red long notes so that they correctly draw the long note body in
between two red notes, previously the long note body was replaced with
standard red notes.

Also color the border of the highway to make it more visible, change the
color of the score panel's background to black, and make the highway
and score panel's height larger.