Add time measurement to animation thread.
Add time measurement to the animation thread so that it can make sure
that it refreshes once every eight milliseconds.

Before, the sleep method being used to make the animation thread iterate
once per 8 milliseconds was using a constant value of 8. Now, it
measures the time taken to draw the HighwayPanel and calculates how long
to wait to ensure that one iteration of the loop takes exactly 8
milliseconds.